v2_services/agent_64_software_test_runner/requirements.txt
fastapi>=0.110
uvicorn>=0.27
pillow>=10.0
imagehash>=4.3
psycopg2-binary>=2.9
httpx>=0.26
minio>=7.0
